Add nogil declaration to _run_paradiso function
This releases the Python GIL around the low level paradiso solver call.
Hopefully this allows this library and dependent libraries (like SimPEG)
to operate more smoothly in a multi-threaded environment.

This should be safe from a Python object perspective because no Python
objects are created/destroyed/inc-reffed during this call.  It could
still be that multiple concurrent calls of this function on the same
data cause undesried results, but hopefully downstream users can apply
concurrency primitives (like locks) as needed.